:white_check_mark:  create, update 오버로딩 추가, Image server 연결 수정

+++ storeServer/store/src/main/java/com/store/Service/StoreServiceImpl.java
@@ -1,8 +1,11 @@
 package com.store.Service;
 
+import com.store.Controller.StoreController;
 import com.store.DAO.StoreDAO;
 import com.store.DTO.*;
 import com.store.Entity.StoreEntity;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cloud.client.ServiceInstance;
 import org.springframework.cloud.client.discovery.DiscoveryClient;
@@ -14,6 +17,8 @@ import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.multipart.MultipartFile;
+
+import javax.swing.text.StyledEditorKit;
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -21,16 +26,28 @@ import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 @Service
 public class StoreServiceImpl implements StoreService {
     private final StoreDAO storeDAO;
     private final DiscoveryClient discoveryClient;
+    private final static Logger logger = LoggerFactory.getLogger(StoreController.class);
 
     public StoreServiceImpl(@Autowired StoreDAO storeDAO,
                             @Autowired DiscoveryClient discoveryClient) {
         this.storeDAO = storeDAO;
         this.discoveryClient = discoveryClient;
+    }
+
+    @Override
+    public SimpleResponseDTO createStore(StoreDTO storeDTO) {
+        StoreEntity storeEntity = toEntity(storeDTO);
+
+        Map<String, Object> resultMap = storeDAO.createStore(storeEntity);
+        SimpleResponseDTO simpleResponseDTO = toSimpleResponseDTO(resultMap);
+        return simpleResponseDTO;
     }
 
     @Override
@@ -50,44 +67,6 @@ public class StoreServiceImpl implements StoreService {
         }
     }
 
-    public String imageUpload(MultipartFile image) throws URISyntaxException, IOException {
-        ByteArrayResource body = new ByteArrayResource(image.getBytes()) {
-            @Override
-            public String getFilename() {
-                return image.getOriginalFilename();
-            }
-        };
-
-        try {
-            ServiceInstance imageService = discoveryClient.getInstances("IMAGE-SERVICE").get(0);
-            RestTemplate restTemplate = new RestTemplate();
-            HttpHeaders headers = new HttpHeaders();
-            HttpEntity<?> http = new HttpEntity<>(headers);
-
-            MultiValueMap<String, Object> bodyMap = new LinkedMultiValueMap<>();
-            bodyMap.add("images", body);
-            headers.setContentType(MediaType.MULTIPART_FORM_DATA);
-            http = new HttpEntity<>(bodyMap, headers);
-
-//            URI uri = new URI(imageService.getUri() + "/image/upload");
-            URI uri = new URI("https://port-0-image-jvpb2mloft5vlw.sel5.cloudtype.app/image/upload");
-            ResponseEntity response = restTemplate.exchange(uri, HttpMethod.POST, http, LinkedHashMap.class);
-
-
-            if (response.getStatusCode().is2xxSuccessful()) {
-                LinkedHashMap responseBody = (LinkedHashMap) response.getBody();
-                List<String> images = (List) responseBody.get("images");
-                String imageUri = (String) images.get(0);
-
-                return imageUri;
-            }
-        } catch (HttpClientErrorException e) {
-            return "Failed to upload image";
-        }
-
-        return "Failed to upload image";
-    }
-    
     @Override
     public ResponseDTO readStoreById(Long storeId) {
         Map<String, Object> resultMap = storeDAO.readStoreById(storeId);
@@ -153,15 +132,42 @@ public class StoreServiceImpl implements StoreService {
     }
 
     @Override
-    public SimpleResponseDTO updateStoreByNameAndSeller(StoreDTO storeDTO){
+    public SimpleResponseDTO updateStoreByNameAndSeller(StoreDTO storeDTO) {
         StoreEntity storeEntity = toEntity(storeDTO);
         Map<String, Object> existMap = storeDAO.existStoreByNameAndSeller(storeEntity.getStoreName(), storeEntity.getStoreSeller());
         SimpleResponseDTO simpleResponseDTO;
 
-        if(existMap.get("result").equals("success")){
+        if (existMap.get("result").equals("success")) {
             Map<String, Object> resultMap = storeDAO.updateStoreByNameAndSeller(storeEntity);
             simpleResponseDTO = toSimpleResponseDTO(resultMap);
-        } else{
+        } else {
+            simpleResponseDTO = toSimpleResponseDTO(existMap);
+        }
+
+        return simpleResponseDTO;
+    }
+
+    @Override
+    public SimpleResponseDTO updateStoreByNameAndSeller(StoreDTO storeDTO, MultipartFile image) {
+        StoreEntity storeEntity = toEntity(storeDTO);
+        Map<String, Object> existMap = storeDAO.existStoreByNameAndSeller(storeEntity.getStoreName(), storeEntity.getStoreSeller());
+        SimpleResponseDTO simpleResponseDTO;
+
+        if (existMap.get("result").equals("success")) {
+            try {
+                String oldImage = storeDAO.readStoreImageByStoreNameAndStoreSeller(storeEntity.getStoreName(), storeEntity.getStoreSeller());
+                String imageKey = findImageKey(oldImage);
+                String imageUri = imageUpdate(image, imageKey);
+                storeEntity.setStoreImage(imageUri);
+                Map<String, Object> resultMap = storeDAO.updateStoreByNameAndSeller(storeEntity);
+                simpleResponseDTO = toSimpleResponseDTO(resultMap);
+                return simpleResponseDTO;
+            } catch (URISyntaxException | IOException e) {
+                e.printStackTrace();
+                SimpleResponseDTO errorResponse = new SimpleResponseDTO("Failed to update image");
+                return errorResponse;
+            }
+        } else {
             simpleResponseDTO = toSimpleResponseDTO(existMap);
         }
 
@@ -174,10 +180,37 @@ public class StoreServiceImpl implements StoreService {
         Map<String, Object> existMap = storeDAO.existStoreById(storeDTO.getStoreId());
         SimpleResponseDTO simpleResponseDTO;
 
-        if(existMap.get("result").equals("success")){
+        if (existMap.get("result").equals("success")) {
             Map<String, Object> resultMap = storeDAO.updateStoreById(storeEntity);
             simpleResponseDTO = toSimpleResponseDTO(resultMap);
-        } else{
+        } else {
+            simpleResponseDTO = toSimpleResponseDTO(existMap);
+        }
+
+        return simpleResponseDTO;
+    }
+
+    @Override
+    public SimpleResponseDTO updateStoreById(StoreDTO storeDTO, MultipartFile image) {
+        StoreEntity storeEntity = toEntity(storeDTO);
+        Map<String, Object> existMap = storeDAO.existStoreById(storeEntity.getStoreId());
+        SimpleResponseDTO simpleResponseDTO;
+
+        if (existMap.get("result").equals("success")) {
+            try {
+                String oldImage = storeDAO.readStoreImageByStoreId(storeEntity.getStoreId());
+                String imageKey = findImageKey(oldImage);
+                String imageUri = imageUpdate(image, imageKey);
+                storeEntity.setStoreImage(imageUri);
+                Map<String, Object> resultMap = storeDAO.updateStoreById(storeEntity);
+                simpleResponseDTO = toSimpleResponseDTO(resultMap);
+                return simpleResponseDTO;
+            } catch (URISyntaxException | IOException e) {
+                e.printStackTrace();
+                SimpleResponseDTO errorResponse = new SimpleResponseDTO("Failed to update image");
+                return errorResponse;
+            }
+        } else {
             simpleResponseDTO = toSimpleResponseDTO(existMap);
         }
 
@@ -186,10 +219,24 @@ public class StoreServiceImpl implements StoreService {
 
     @Override
     public SimpleResponseDTO deleteStore(Long storeId) {
-        Map<String, Object> resultMap = storeDAO.deleteStore(storeId);
+        try {
+            String image = storeDAO.readStoreImageByStoreId(storeId);
+            String imageKey = findImageKey(image);
 
-        SimpleResponseDTO simpleResponseDTO = toSimpleResponseDTO(resultMap);
-        return simpleResponseDTO;
+            if (imageDelete(imageKey)) {
+                Map<String, Object> resultMap = storeDAO.deleteStore(storeId);
+
+                SimpleResponseDTO simpleResponseDTO = toSimpleResponseDTO(resultMap);
+                return simpleResponseDTO;
+            } else {
+                SimpleResponseDTO errorResponse = new SimpleResponseDTO("Failed to Delete image");
+                return errorResponse;
+            }
+        } catch (URISyntaxException | IOException e) {
+            e.printStackTrace();
+            SimpleResponseDTO errorResponse = new SimpleResponseDTO("Failed to Delete image");
+            return errorResponse;
+        }
     }
 
     @Override
@@ -257,7 +304,7 @@ public class StoreServiceImpl implements StoreService {
                     .result((String) resultMap.get("result"))
                     .data(readResponseDTOS)
                     .build();
-        } else{
+        } else {
             responseDTO = ResponseDTO.builder()
                     .result((String) resultMap.get("result"))
                     .data(null)
@@ -287,10 +334,10 @@ public class StoreServiceImpl implements StoreService {
                     .result((String) resultMap.get("result"))
                     .data(readResponseForMapDTO)
                     .build();
-        } else if(resultMap.containsKey("dataList")){
+        } else if (resultMap.containsKey("dataList")) {
             List<StoreEntity> storeEntities = (List<StoreEntity>) resultMap.get("dataList");
 
-            if(storeEntities.isEmpty()){
+            if (storeEntities.isEmpty()) {
                 responseDTO = ResponseDTO.builder()
                         .result((String) resultMap.get("result"))
                         .data(null)
@@ -336,5 +383,137 @@ public class StoreServiceImpl implements StoreService {
                 .build();
 
         return simpleResponseDTO;
+    }
+
+    public String imageUpload(MultipartFile image) throws URISyntaxException, IOException {
+        ByteArrayResource body = new ByteArrayResource(image.getBytes()) {
+            @Override
+            public String getFilename() {
+                return image.getOriginalFilename();
+            }
+        };
+
+        try {
+            ServiceInstance imageService = discoveryClient.getInstances("IMAGE-SERVICE").get(0);
+            RestTemplate restTemplate = new RestTemplate();
+            HttpHeaders headers = new HttpHeaders();
+            HttpEntity<?> http = new HttpEntity<>(headers);
+
+            MultiValueMap<String, Object> bodyMap = new LinkedMultiValueMap<>();
+            bodyMap.add("images", body);
+            headers.setContentType(MediaType.MULTIPART_FORM_DATA);
+            http = new HttpEntity<>(bodyMap, headers);
+
+//            URI uri = new URI(imageService.getUri() + "/image/upload");
+            URI uri = new URI("https://port-0-image-jvpb2mloft5vlw.sel5.cloudtype.app/image/upload");
+            ResponseEntity response = restTemplate.exchange(uri, HttpMethod.POST, http, LinkedHashMap.class);
+
+
+            if (response.getStatusCode().is2xxSuccessful()) {
+                LinkedHashMap responseBody = (LinkedHashMap) response.getBody();
+                List<String> images = (List) responseBody.get("images");
+                String imageUri = (String) images.get(0);
+
+                return imageUri;
+            }
+        } catch (HttpClientErrorException e) {
+            return "Failed to upload image";
+        }
+
+        return "Failed to upload image";
+    }
+
+    public String imageUpdate(MultipartFile image, String imageKey) throws URISyntaxException, IOException {
+        ByteArrayResource body = new ByteArrayResource(image.getBytes()) {
+            @Override
+            public String getFilename() {
+                return image.getOriginalFilename();
+            }
+        };
+
+        try {
+            ServiceInstance imageService = discoveryClient.getInstances("IMAGE-SERVICE").get(0);
+            RestTemplate restTemplate = new RestTemplate();
+            HttpHeaders headers = new HttpHeaders();
+            HttpEntity<?> http = new HttpEntity<>(headers);
+
+            MultiValueMap<String, Object> bodyMap = new LinkedMultiValueMap<>();
+            bodyMap.add("image", body);
+            headers.setContentType(MediaType.MULTIPART_FORM_DATA);
+            http = new HttpEntity<>(bodyMap, headers);
+            ResponseEntity response;
+
+//            URI uri = new URI(imageService.getUri() + "/image/upload");
+            if (imageKey != null) {
+                URI uri = new URI("https://port-0-image-jvpb2mloft5vlw.sel5.cloudtype.app/image/" + imageKey);
+                response = restTemplate.exchange(uri, HttpMethod.PUT, http, String.class);
+
+                if (response.getStatusCode().is2xxSuccessful()) {
+                    String imageUri = (String) response.getBody();
+
+                    return imageUri;
+                }
+            } else {
+                URI uri = new URI("https://port-0-image-jvpb2mloft5vlw.sel5.cloudtype.app/image/upload");
+                response = restTemplate.exchange(uri, HttpMethod.POST, http, LinkedHashMap.class);
+                logger.info("ImageServer POST method");
+
+                if (response.getStatusCode().is2xxSuccessful()) {
+                    LinkedHashMap responseBody = (LinkedHashMap) response.getBody();
+                    List<String> images = (List) responseBody.get("images");
+                    String imageUri = (String) images.get(0);
+
+                    return imageUri;
+                }
+            }
+
+        } catch (HttpClientErrorException e) {
+            return "Failed to upload image";
+        }
+
+        return "Failed to upload image";
+    }
+
+    public Boolean imageDelete(String imageKey) throws URISyntaxException, IOException {
+        try {
+            ServiceInstance imageService = discoveryClient.getInstances("IMAGE-SERVICE").get(0);
+            RestTemplate restTemplate = new RestTemplate();
+            HttpHeaders headers = new HttpHeaders();
+            HttpEntity<?> http = new HttpEntity<>(headers);
+
+            http = new HttpEntity<>(headers);
+            ResponseEntity response;
+
+//            URI uri = new URI(imageService.getUri() + "/image/upload");
+            if (imageKey != null) {
+                URI uri = new URI("https://port-0-image-jvpb2mloft5vlw.sel5.cloudtype.app/image/" + imageKey);
+                response = restTemplate.exchange(uri, HttpMethod.DELETE, http, Boolean.class);
+                logger.info("ImageServer Delete method");
+                return (Boolean) response.getBody();
+            } else {
+                return true;
+            }
+        } catch (HttpClientErrorException e) {
+            return false;
+        }
+    }
+
+    public String findImageKey(String url) {
+        // 정규 표현식 패턴을 정의합니다.
+        Pattern pattern = Pattern.compile("/image/download/(\\d+)");
+
+        // 패턴과 URL을 매칭시킵니다.
+        Matcher matcher = pattern.matcher(url);
+
+        String imageKey;
+        // 매칭된 부분을 찾습니다.
+        if (matcher.find()) {
+            // 숫자를 문자열로 추출합니다.
+            imageKey = matcher.group(1);
+        } else {
+            imageKey = null;
+        }
+        logger.info("ImageKey: " + imageKey);
+        return imageKey;
     }
 }
